requirements.txt
Django==4.1.5
environs==9.5.0
Pillow==9.4.0
django-debug-toolbar==3.8.1
django-tinymce==3.5.0
django-phonenumber-field[phonenumberslite]
